Move changedFilesMap into a separate module
