+++ src/pfe/file-watcher/server/src/projects/libertyProject.ts
@@ -26,8 +26,8 @@ import * as locale from "../utils/locale";
 import * as logger from "../utils/logger";
 import * as logHelper from "./logHelper";
 import * as kubeutil from "../utils/kubeutil";
-import * as projectEventsController from "../controllers/projectEventsController";
 import { ProcessResult } from "../utils/processManager";
+import { IFileChangeEvent } from "../utils/fileChanges";
 
 
 const readFileAsync = promisify(fs.readFile);
@@ -111,11 +111,11 @@ export function create(operation: Operation): void {
  * @description Update operation for a liberty project.
  *
  * @param operation <Required | Operation> - The update operation.
- * @param changedFiles <Optional | projectEventsController.IFileChangeEvent[]> - The file changed event array.
+ * @param changedFiles <Optional | IFileChangeEvent[]> - The file changed event array.
  *
  * @returns void
  */
-export function update(operation: Operation, changedFiles?: projectEventsController.IFileChangeEvent[]): void {
+export function update(operation: Operation, changedFiles?: IFileChangeEvent[]): void {
     projectUtil.containerUpdate(operation, "/file-watcher/scripts/liberty-container.sh", "update");
 }
